refactor(ksql): adds docs for sql and refactor the sql driver code
